% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accessor_functions.R
\name{Accessors}
\alias{Accessors}
\alias{get_hyperparam}
\alias{get_intercept}
\title{Accessor functions for \code{ipriorMod} objects.}
\usage{
get_hyperparam(object)

get_intercept(object, by.class = FALSE)
}
\arguments{
\item{object}{An \code{ipriorMod} object.}
}
\description{
Accessor functions for \code{ipriorMod} objects.
}
\section{Functions}{
\itemize{
\item \code{get_hyperparam}: Obtain all of the hyperparameters.

\item \code{get_intercept}: Obtain the intercept.
}}

