% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{iprobit-package}
\alias{iprobit-package}
\title{iprobit: Binary and Multinomial Probit Regression with I-priors.}
\description{
Fit I-prior probit models to binary and multinomial data for classification
or inference using fitted probabilities.
}
\section{iprobit functions}{

iprobit()
}

