% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.iprobitMod}
\alias{logLik.iprobitMod}
\title{Extract the variational lower bound}
\usage{
\method{logLik}{iprobitMod}(x)
}
\arguments{
\item{x}{an object of class \code{ipriorProbit}.}
}
\value{
The variational lower bound.
}
\description{
Extract the variational lower bound
}
