% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ipriorProbit}
\alias{plot.ipriorProbit}
\title{Plot for \code{ipriorProbit} objects.}
\usage{
\method{plot}{ipriorProbit}(x, niter.plot = NULL, levels = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot for \code{ipriorProbit} objects.
}
