% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.iprobitMod}
\alias{logLik.iprobitMod}
\title{Extract the variational lower bound}
\usage{
\method{logLik}{iprobitMod}(object, theta = NULL, silent = TRUE,
  stop.crit = 0.001, alpha = get_alpha(object), ...)
}
\arguments{
\item{object}{An object of class \code{ipriorProbit}.}

\item{...}{This is not used here.}
}
\value{
The variational lower bound.
}
\description{
Extract the variational lower bound
}
